Move to a transaction-based caching model.
Every cache operation is organized into _transactions_, which are
represented through the `Transaction` trait.

- Use `usize` instead of `u64` for representing sectors and sector size.
- Add the constraint that the sector size must be a power of two.
- Detect metacycles in the cache dependency graph.

+++ src/io/disk.rs
@@ -1,5 +1,5 @@
 /// A disk sector number.
-type Sector = u64;
+type Sector = usize;
 
 /// A disk I/O error.
 enum Error {
